fix timestamp parsing for events
