+++ graylog2-web-interface/src/components/common/EntityFilters/helpers/timeRange.test.ts
@@ -16,7 +16,11 @@
  */
 import type { DateTime } from 'util/DateTime';
 
-import { timeRangePickerFormValuesToFilterValue, timeRangeTitle } from './timeRange';
+import {
+  filterValueToTimeRangePickerFormValues,
+  timeRangePickerFormValuesToFilterValue,
+  timeRangeTitle,
+} from './timeRange';
 
 const formatTime = (dateTime: DateTime) => `formatted:${dateTime}`;
 
@@ -155,6 +159,116 @@ describe('timeRange helpers', () => {
       ).toEqual({
         title: 'formatted:2024-05-01T12:10:41.000+00:00 - formatted:2026-01-11T12:15:46.624+00:00',
         value: 'absolute@2024-05-01T12:10:41.000+00:00><2026-01-11T12:15:46.624+00:00',
+      });
+    });
+  });
+
+  describe('filterValueToTimeRangePickerFormValues', () => {
+    it('converts relative start-only filter values', () => {
+      expect(filterValueToTimeRangePickerFormValues('relative@300', formatTime)).toEqual({
+        activeTab: 'relative',
+        timeRangeTabs: {
+          relative: {
+            type: 'relative',
+            from: {
+              value: 5,
+              unit: 'minutes',
+              isAllTime: false,
+            },
+            to: {
+              value: 0,
+              unit: 'seconds',
+              isAllTime: true,
+            },
+          },
+        },
+      });
+    });
+
+    it('converts relative range filter values', () => {
+      expect(filterValueToTimeRangePickerFormValues('relative@300><200', formatTime)).toEqual({
+        activeTab: 'relative',
+        timeRangeTabs: {
+          relative: {
+            type: 'relative',
+            from: {
+              value: 5,
+              unit: 'minutes',
+              isAllTime: false,
+            },
+            to: {
+              value: 200,
+              unit: 'seconds',
+              isAllTime: false,
+            },
+          },
+        },
+      });
+    });
+
+    it('converts keyword filter values', () => {
+      expect(filterValueToTimeRangePickerFormValues('keyword@Last+five+minutes', formatTime)).toEqual({
+        activeTab: 'keyword',
+        timeRangeTabs: {
+          keyword: {
+            type: 'keyword',
+            keyword: 'Last five minutes',
+          },
+        },
+      });
+    });
+
+    it('converts absolute filter values to user timezone form values', () => {
+      expect(
+        filterValueToTimeRangePickerFormValues(
+          'absolute@2024-05-01T12%3A10%3A41.000%2B00%3A00><2026-05-11T12%3A15%3A46.624%2B00%3A00',
+          formatTime,
+        ),
+      ).toEqual({
+        activeTab: 'absolute',
+        timeRangeTabs: {
+          absolute: {
+            type: 'absolute',
+            from: 'formatted:2024-05-01T12:10:41.000+00:00',
+            to: 'formatted:2026-05-11T12:15:46.624+00:00',
+          },
+        },
+      });
+    });
+
+    it('converts legacy untyped absolute filter values', () => {
+      expect(
+        filterValueToTimeRangePickerFormValues(
+          '2024-05-01T12%3A10%3A41.000%2B00%3A00><2026-05-11T12%3A15%3A46.624%2B00%3A00',
+          formatTime,
+        ),
+      ).toEqual({
+        activeTab: 'absolute',
+        timeRangeTabs: {
+          absolute: {
+            type: 'absolute',
+            from: 'formatted:2024-05-01T12:10:41.000+00:00',
+            to: 'formatted:2026-05-11T12:15:46.624+00:00',
+          },
+        },
+      });
+    });
+
+    it('keeps raw timezone offsets in legacy absolute filter values', () => {
+      expect(
+        filterValueToTimeRangePickerFormValues(
+          '2024-05-01T12:10:41.000+00:00><2026-05-11T12:15:46.624+00:00',
+          formatTime,
+        ),
+      ).toEqual({
+        activeTab: 'absolute',
+        timeRangeTabs: {
+          absolute: {
+            type: 'absolute',
+            from: 'formatted:2024-05-01T12:10:41.000+00:00',
+            to: 'formatted:2026-05-11T12:15:46.624+00:00',
+          },
+        },
       });
     });
   });
